# lib/swappiq-api.js
// SwappiQ API implementation for Vercel
const crypto = require('crypto');

// In-memory stores
const orders = new Map();
const orderBook = { buy: [], sell: [] };
const users = new Map();
const balances = new Map();
const settlements = new Map();
const disputes = new Map();
const notifications = [];
const marketMakerApplications = new Map();

// Mock token data
const tokens = [
  { id: 1, symbol: 'ETH', name: 'Ethereum', address: '0x0000000000000000000000000000000000000000', decimals: 18, chainId: 1, logoURI: 'https://ethereum.org/icon.png' },
  { id: 2, symbol: 'USDT', name: 'Tether USD', address: '0xdac17f958d2ee523a2206206994597c13d831ec7', decimals: 6, chainId: 1, logoURI: 'https://tether.to/icon.png' },
  { id: 3, symbol: 'USDC', name: 'USD Coin', address: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48', decimals: 6, chainId: 1, logoURI: 'https://centre.io/icon.png' },
  { id: 4, symbol: 'DAI', name: 'Dai Stablecoin', address: '0x6b175474e89094c44da98b954eedeac495271d0f', decimals: 18, chainId: 1, logoURI: 'https://makerdao.com/icon.png' },
  { id: 5, symbol: 'WBTC', name: 'Wrapped Bitcoin', address: '0x2260fac5e5542a773aa44fbcfedf7c193bc2c599', decimals: 8, chainId: 1, logoURI: 'https://wbtc.network/icon.png' }
];

// Initialize balances
balances.set('user1', { ETH: 10, USDT: 10000, USDC: 10000 });
balances.set('user2', { ETH: 5, USDT: 5000, USDC: 5000 });

// Utility functions
function generateOrderId() {
  return 'order_' + Date.now() + '_' + Math.random().toString(36).substr(2, 9);
}

function generateJWT(userId) {
  const header = Buffer.from(JSON.stringify({ typ: 'JWT', alg: 'HS256' })).toString('base64');
  const payload = Buffer.from(JSON.stringify({ userId, exp: Date.now() + 86400000 })).toString('base64');
  const signature = crypto.createHmac('sha256', 'secret').update(header + '.' + payload).digest('base64');
  return header + '.' + payload + '.' + signature;
}

function verifyJWT(token) {
  try {
    const [header, payload, signature] = token.split('.');
    if (!header || !payload || !signature) return null;
    const expectedSignature = crypto.createHmac('sha256', 'secret').update(header + '.' + payload).digest('base64');
    if (signature !== expectedSignature) return null;
    const data = JSON.parse(Buffer.from(payload, 'base64').toString());
    if (data.exp < Date.now()) return null;
    return data;
  } catch (e) {
    return null;
  }
}

// API Routes
const routes = {
  // Health endpoints
  'GET /api/health': (req, res) => {
    res.json({ status: 'healthy', timestamp: new Date().toISOString() });
  },
  
  'GET /api/health/detailed': (req, res) => {
    res.json({
      status: 'healthy',
      services: {
        database: 'connected',
        redis: 'connected',
        blockchain: 'connected',
        marketMaker: 'active'
      },
      metrics: {
        activeOrders: orders.size,
        totalSettlements: settlements.size,
        activeDisputes: disputes.size
      }
    });
  },

  // Token endpoints
  'GET /api/tokens/comprehensive': (req, res) => {
    res.json({ tokens, total: tokens.length });
  },

  // Quote endpoints
  'POST /api/quote': async (req, res) => {
    const { sellToken, buyToken, sellAmount } = req.body;
    const rate = sellToken === 'ETH' ? 2000 : 1;
    const buyAmount = (parseFloat(sellAmount) * rate).toString();
    
    res.json({
      sellToken,
      buyToken,
      sellAmount,
      buyAmount,
      price: rate.toString(),
      estimatedGas: '150000',
      sources: [{ name: 'SwappiQ', proportion: '1' }]
    });
  },

  // Order endpoints
  'POST /api/submitOrder': async (req, res) => {
    const auth = req.headers.authorization;
    if (!auth || !auth.startsWith('Bearer ') || !verifyJWT(auth.replace('Bearer ', ''))) {
      return res.status(401).json({ error: 'Unauthorized' });
    }

    const order = {
      id: generateOrderId(),
      ...req.body,
      status: 'pending',
      timestamp: new Date().toISOString()
    };
    
    orders.set(order.id, order);
    orderBook[order.side].push(order);
    
    // Simulate matching
    setTimeout(() => {
      order.status = 'filled';
      notifications.push({
        userId: order.userId,
        type: 'order_filled',
        orderId: order.id,
        timestamp: new Date().toISOString()
      });
    }, 2000);
    
    res.json({ orderId: order.id, status: 'pending' });
  },

  'GET /api/orders/history': (req, res) => {
    const auth = req.headers.authorization;
    if (!auth || !auth.startsWith('Bearer ')) {
      return res.status(401).json({ error: 'Unauthorized' });
    }
    const jwt = verifyJWT(auth.replace('Bearer ', ''));
    if (!jwt) {
      return res.status(401).json({ error: 'Unauthorized' });
    }

    const userOrders = Array.from(orders.values())
      .filter(o => o.userId === jwt.userId)
      .sort((a, b) => new Date(b.timestamp) - new Date(a.timestamp));
    
    res.json({ orders: userOrders });
  },

  'GET /api/orders/:orderId': (req, res) => {
    const order = orders.get(req.params.orderId);
    if (!order) {
      return res.status(404).json({ error: 'Order not found' });
    }
    res.json(order);
  },

  'GET /api/orderbook': (req, res) => {
    res.json({
      buy: orderBook.buy.slice(-50),
      sell: orderBook.sell.slice(-50),
      timestamp: new Date().toISOString()
    });
  },

  // Settlement endpoints
  'POST /api/settlement/initiate': (req, res) => {
    const settlementId = 'settlement_' + Date.now();
    const settlement = {
      id: settlementId,
      ...req.body,
      status: 'initiated',
      timestamp: new Date().toISOString()
    };
    settlements.set(settlementId, settlement);
    res.json({ settlementId, status: 'initiated' });
  },

  'GET /api/orders/:orderId/settlement-proof': (req, res) => {
    const order = orders.get(req.params.orderId);
    if (!order || order.status !== 'filled') {
      return res.status(404).json({ error: 'No settlement proof available' });
    }
    
    const proof = {
      orderId: order.id,
      merkleRoot: crypto.randomBytes(32).toString('hex'),
      proofPath: [
        crypto.randomBytes(32).toString('hex'),
        crypto.randomBytes(32).toString('hex')
      ],
      blockNumber: Math.floor(Math.random() * 1000000) + 15000000,
      timestamp: order.timestamp
    };
    
    res.json(proof);
  },

  // Dispute endpoints
  'POST /api/disputes': (req, res) => {
    const dispute = {
      id: 'dispute_' + Date.now(),
      ...req.body,
      status: 'open',
      timestamp: new Date().toISOString()
    };
    disputes.set(dispute.id, dispute);
    res.json({ disputeId: dispute.id, status: 'open' });
  },

  'POST /api/disputes/settle': (req, res) => {
    const { disputeId, resolution } = req.body;
    const dispute = disputes.get(disputeId);
    if (!dispute) {
      return res.status(404).json({ error: 'Dispute not found' });
    }
    dispute.status = 'resolved';
    dispute.resolution = resolution;
    res.json({ disputeId, status: 'resolved' });
  },

  // Market maker endpoints
  'POST /api/market-maker/apply': (req, res) => {
    const application = {
      id: 'mm_app_' + Date.now(),
      ...req.body,
      status: 'pending',
      timestamp: new Date().toISOString()
    };
    marketMakerApplications.set(application.id, application);
    res.json({ applicationId: application.id, status: 'pending' });
  },

  'GET /api/competition/leaderboard': (req, res) => {
    const leaderboard = [
      { rank: 1, marketMaker: 'MM Alpha', score: 95.5, volume: '1000000' },
      { rank: 2, marketMaker: 'MM Beta', score: 92.3, volume: '850000' },
      { rank: 3, marketMaker: 'MM Gamma', score: 88.7, volume: '720000' }
    ];
    res.json({ leaderboard, lastUpdated: new Date().toISOString() });
  },

  // Analytics endpoints
  'GET /api/analytics/profits': (req, res) => {
    res.json({
      daily: '1250.50',
      weekly: '8750.25',
      monthly: '35200.00',
      currency: 'USDT'
    });
  },

  'GET /api/revenue/status': (req, res) => {
    res.json({
      totalRevenue: '125000.00',
      pendingSettlements: '15000.00',
      completedSettlements: '110000.00',
      currency: 'USDT'
    });
  },

  // Notification endpoints
  'GET /api/notifications/user/:userId': (req, res) => {
    const userNotifications = notifications.filter(n => n.userId === req.params.userId);
    res.json({ notifications: userNotifications });
  },

  'GET /api/ws/health': (req, res) => {
    res.json({ status: 'active', connections: 0 });
  }
};

// Main handler
export function swappiqStandalone(req, res) {
  // Extract path from Next.js query
  const path = '/api/' + (req.query.path ? req.query.path.join('/') : '');
  const method = req.method;
  
  // Helper functions
  res.json = (data) => {
    res.setHeader('Content-Type', 'application/json');
    res.status(res.statusCode || 200).send(JSON.stringify(data));
  };
  
  res.status = (code) => {
    res.statusCode = code;
    return res;
  };

  // Find matching route
  let routeKey = `${method} ${path}`;
  if (routes[routeKey]) {
    req.params = {};
    return routes[routeKey](req, res);
  }

  // Check for parameterized routes
  for (const [key, handler] of Object.entries(routes)) {
    const [routeMethod, routePattern] = key.split(' ');
    if (routeMethod !== method) continue;
    
    const regex = new RegExp('^' + routePattern.replace(/:(\w+)/g, '(?<$1>[^/]+)') + '$');
    const match = path.match(regex);
    
    if (match) {
      req.params = match.groups || {};
      return handler(req, res);
    }
  }

  // 404
  res.status(404).json({ error: 'Not found' });
}